+++ code/handmade.h
@@ -36,7 +36,7 @@
 #define Gigabytes(Value) (Megabytes(Value) * 1024)
 #define Terabytes(Value) (Gigabytes(Value) * 1024)
 
-#define ArrayCount(Array) (sizeof(Array) / sizeof((Array)[0]))
+#define ArrayCount(Array) static_cast<int>(sizeof(Array) / sizeof((Array)[0]))
 // TODO: swap, min, max ... macros???
 
 inline uint32 SafeTruncateUInt64(uint64 Value) {
